Finished image carousel and function where each scroll's images are added via objects
